refactor: widgetSet rect
